Updated ID type to BIGINT

+++ Backend/database/schema.sql
@@ -2,7 +2,7 @@ BEGIN;
 
 -- Create Users Table
 CREATE TABLE users (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     username TEXT UNIQUE NOT NULL,
     password TEXT NOT NULL,
 
@@ -14,8 +14,8 @@ CREATE TABLE users (
 
 -- Create Origami Table
 CREATE TABLE origami (
-    id SERIAL PRIMARY KEY,
-    user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    user_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     is_public BOOLEAN NOT NULL DEFAULT FALSE,
     ratings DOUBLE PRECISION NOT NULL DEFAULT 0.0,
 
@@ -27,7 +27,7 @@ CREATE TABLE origami (
 
 -- Create Step Type Table
 CREATE TABLE step_type (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     step_type TEXT UNIQUE NOT NULL,
 
     created_by TEXT DEFAULT NULL,
@@ -38,9 +38,9 @@ CREATE TABLE step_type (
 
 -- Create Step Table
 CREATE TABLE step (
-    id SERIAL PRIMARY KEY,
-    origami_id INTEGER NOT NULL REFERENCES origami(id) ON DELETE CASCADE,
-    step_type_id INTEGER NOT NULL REFERENCES step_type(id) ON DELETE RESTRICT,
+    id BIGSERIAL PRIMARY KEY,
+    origami_id BIGINT NOT NULL REFERENCES origami(id) ON DELETE CASCADE,
+    step_type_id BIGINT NOT NULL REFERENCES step_type(id) ON DELETE RESTRICT,
 
     created_by TEXT DEFAULT NULL,
     updated_by TEXT DEFAULT NULL,
@@ -50,10 +50,10 @@ CREATE TABLE step (
 
 -- Create Face Table
 CREATE TABLE face (
-    id SERIAL PRIMARY KEY,
-    origami_id INTEGER NOT NULL REFERENCES origami(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    origami_id BIGINT NOT NULL REFERENCES origami(id) ON DELETE CASCADE,
     face_id_in_origami INTEGER NOT NULL,
-    step_id INTEGER REFERENCES step(id) ON DELETE CASCADE,
+    step_id BIGINT REFERENCES step(id) ON DELETE CASCADE,
     is_deleted BOOLEAN NOT NULL DEFAULT FALSE,
 
     created_by TEXT DEFAULT NULL,
@@ -64,8 +64,8 @@ CREATE TABLE face (
 
 -- Create Fold Step Table
 CREATE TABLE fold_step (
-    step_id INTEGER PRIMARY KEY REFERENCES step(id) ON DELETE CASCADE,
-    anchored_face_id INTEGER REFERENCES face(id) ON DELETE SET NULL,
+    step_id BIGINT PRIMARY KEY REFERENCES step(id) ON DELETE CASCADE,
+    anchored_face_id BIGINT REFERENCES face(id) ON DELETE SET NULL,
 
     created_by TEXT DEFAULT NULL,
     updated_by TEXT DEFAULT NULL,
@@ -75,8 +75,8 @@ CREATE TABLE fold_step (
 
 -- Create Vertex Table
 CREATE TABLE vertex (
-    id SERIAL PRIMARY KEY,
-    face_id INTEGER NOT NULL REFERENCES face(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    face_id BIGINT NOT NULL REFERENCES face(id) ON DELETE CASCADE,
     x_pos DOUBLE PRECISION NOT NULL,
     y_pos DOUBLE PRECISION NOT NULL,
 
@@ -88,10 +88,10 @@ CREATE TABLE vertex (
 
 -- Create Edge Table
 CREATE TABLE edge (
-    id SERIAL PRIMARY KEY,
-    origami_id INTEGER NOT NULL REFERENCES origami(id) ON DELETE CASCADE,
-    face_1_id INTEGER REFERENCES face(id) ON DELETE SET NULL,
-    face_2_id INTEGER REFERENCES face(id) ON DELETE SET NULL,
+    id BIGSERIAL PRIMARY KEY,
+    origami_id BIGINT NOT NULL REFERENCES origami(id) ON DELETE CASCADE,
+    face_1_id BIGINT REFERENCES face(id) ON DELETE SET NULL,
+    face_2_id BIGINT REFERENCES face(id) ON DELETE SET NULL,
     angle DOUBLE PRECISION NOT NULL,
 
     created_by TEXT DEFAULT NULL,
@@ -102,12 +102,12 @@ CREATE TABLE edge (
 
 -- Create Annotated Point Table
 CREATE TABLE annotated_point (
-    id SERIAL PRIMARY KEY,
-    step_id INTEGER NOT NULL REFERENCES step(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    step_id BIGINT NOT NULL REFERENCES step(id) ON DELETE CASCADE,
     x_pos DOUBLE PRECISION NOT NULL,
     y_pos DOUBLE PRECISION NOT NULL,
-    on_edge_id INTEGER REFERENCES edge(id) ON DELETE SET NULL,
-    vertex_id INTEGER REFERENCES vertex(id) ON DELETE SET NULL,
+    on_edge_id BIGINT REFERENCES edge(id) ON DELETE SET NULL,
+    vertex_id BIGINT REFERENCES vertex(id) ON DELETE SET NULL,
     is_deleted BOOLEAN NOT NULL DEFAULT FALSE,
 
     created_by TEXT DEFAULT NULL,
@@ -118,10 +118,10 @@ CREATE TABLE annotated_point (
 
 -- Create Annotated Line Table
 CREATE TABLE annotated_line (
-    id SERIAL PRIMARY KEY,
-    step_id INTEGER NOT NULL REFERENCES step(id) ON DELETE CASCADE,
-    point_1_id INTEGER NOT NULL REFERENCES annotated_point(id) ON DELETE CASCADE,
-    point_2_id INTEGER NOT NULL REFERENCES annotated_point(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    step_id BIGINT NOT NULL REFERENCES step(id) ON DELETE CASCADE,
+    point_1_id BIGINT NOT NULL REFERENCES annotated_point(id) ON DELETE CASCADE,
+    point_2_id BIGINT NOT NULL REFERENCES annotated_point(id) ON DELETE CASCADE,
     is_deleted BOOLEAN NOT NULL DEFAULT FALSE,
 
     created_by TEXT DEFAULT NULL,
